% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{createAccount}
\alias{createAccount}
\title{SpatialMap Account Creation}
\usage{
createAccount(password = "prompt", email = "prompt")
}
\arguments{
\item{password}{Any password with a minimum of complexity.}

\item{email}{Any email that can receive the confirmation mail.}
}
\value{
Message of success or failure.
}
\description{
Uploads require a user login. The createAccount function allows to create the account within R with valid password and email. In case a spatialMaps account was already created on the platform, no new account is neccesssary.
}
