% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{createColors}
\alias{createColors}
\title{Append colors to the fSet table.}
\usage{
createColors(object)
}
\arguments{
\item{object}{The pRoloc object.}
}
\value{
Appends the color column to the fSet.
}
\description{
A helper function to pass pRoloc conform colors to the fData table.
}
