% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataBackup}
\alias{dataBackup}
\title{SpatialMap database backup function}
\usage{
dataBackup(outputFile = "SpatialMaps.json", secret = NULL)
}
\arguments{
\item{outputFile}{The output file name that should include the .json identifier {string}. File is written into the working directory.}

\item{secret}{The database secret key. Can be optained from options -> project config -> accounts -> database secrets.}
}
\value{
An instance of class \code{pRolocmetadata}.
}
\description{
Backup of all SpatialMaps database entries
}
