% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download}
\alias{download}
\title{Download Datasets from SpatialMaps.}
\usage{
download(dataset)
}
\arguments{
\item{dataset}{The shortlink received from SpatialMaps}
}
\value{
Success or failure message
}
\description{
SpatialMaps download function. The dataset name can be gathered from the SpatialMaps platform.
}
