\name{E14TG2a}
\alias{E14TG2a}
\alias{E14TG2aS1}
\alias{E14TG2aS2}
\alias{E14TG2aR}

\docType{data}

\title{
  Data from a LOPIT experiment on Mouse E14TG2a Embryonic Stem Cells
}

\description{
  This is data from a standard LOPIT experimental design on Mouse E14TG2a
  embryonic stem cells
}

\usage{
data(E14TG2aS1)
data(E14TG2aS2)
data(E14TG2aR)
}

\format{
  The data is an instance of class \code{MSnSet} from package \code{MSnbase}.
}

\details{
  This is a LOPIT experiment. Normalised intensities of proteins from eight
  iTRAQ 8-plex labelled fractions are available for 2 replicates (indexed 1 
  and 2) using stringent and relaxed setting (S and R, respectively). 
}

\source{
  The data was generated by A. Christoforou at the Cambridge Centre for 
  Proteomics.
  \url{http://www.bio.cam.ac.uk/proteomics/}. 
}


\examples{
data(E14TG2aS1)
E14TG2aS1
pData(E14TG2aS1)
head(exprs(E14TG2aS1))
}

\keyword{datasets}
