% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filterKeys}
\alias{filterKeys}
\title{Filter forbidden json keys}
\usage{
filterKeys(x)
}
\arguments{
\item{x}{The column names of the dataset.}
}
\value{
returns correctedNames and reports changes in the naming.
}
\description{
The firebase realtime database probhibts certain synthax relevant column names. This function helps to overcome potential errors and warns the user if one or more column names needed to be renamed.
}
