\name{havugimana2012}
\alias{havugimana2012}
\docType{data}


\title{
  Data from Havugimana et al. 2012
}

\description{

  Data from Havugimana et al. 2012, \emph{A census of human soluble
    protein complexes}. The protein complexes were fractionated by ion
  exchange chromatography, Isoelectric focusing and sucrose density
  gradient centrifugation. Proteins were quantified by spectral
  counting.
  
}

\usage{data("havugimana2012")}



\references{
  Havugimana PC, Hart GT, Nepusz T, Yang H, Turinsky AL, Li Z, Wang PI,
  Boutz DR, Fong V, Phanse S, Babu M, Craig SA, Hu P, Wan C, Vlasblom J,
  Dar VU, Bezginov A, Clark GW, Wu GC, Wodak SJ, Tillier ER, Paccanaro
  A, Marcotte EM, Emili A. A census of human soluble protein
  complexes. Cell. 2012 Aug 31;150(5):1068-81. doi:
  10.1016/j.cell.2012.08.011. PubMed PMID: 22939629; PubMed Central
  PMCID: PMC3477804.
}

\examples{
data(havugimana2012)
experimentData(havugimana2012)
}

\keyword{datasets}
