\name{kirkwood2013}
\alias{kirkwood2013}

\docType{data}

\title{
  Data from Kirkwood et al. 2013.
}

\description{

  Data from Kirkwood et al. 2013, \emph{Characterization of native
  protein complexes and protein isoform variation using
  size-fractionation-based quantitative proteomics}. Protein complexes
  were separated by size exclusion chromatography and proteins were
  quantified by spectral counting.
  
}

\usage{data("kirkwood2013")}


\references{
  Kirkwood KJ, Ahmad Y, Larance M, Lamond AI. Characterization of native
  protein complexes and protein isoform variation using
  size-fractionation-based quantitative proteomics. Mol Cell
  Proteomics. 2013 Dec;12(12):3851-73. doi:
  10.1074/mcp.M113.032367. Epub 2013 Sep 16. PubMed PMID: 24043423;
  PubMed Central PMCID: PMC3861729.
}

\examples{
data(kirkwood2013)
experimentData(kirkwood2013)
}

\keyword{datasets}
