\name{kristensen2012r1}
\alias{kristensen2012r1}
\alias{kristensen2012r2}
\alias{kristensen2012r3}
\alias{kristensen2012}

\docType{data}
\title{
  Data from Kristensen et al. 2012
}

\description{

  Triplicated experimental data from Kristensen et al. 2012, \emph{A
  high-throughput approach for measuring temporal changes in the
  interactome}. Protein complexes were separated by size exclusion
  chromatography and protein were quantified using SILAC.

}

\usage{
data("kristensen2012r1")
data("kristensen2012r2")
data("kristensen2012r3")
}


\references{
  Kristensen AR, Gsponer J, Foster LJ. A high-throughput approach for
  measuring temporal changes in the interactome. Nat Methods. 2012
  Sep;9(9):907-9. doi: 10.1038/nmeth.2131. Epub 2012 Aug 5. PubMed PMID:
  22863883; PubMed Central PMCID: PMC3954081.
}


\examples{
data(kristensen2012r1)
experimentData(kristensen2012r1)
}

\keyword{datasets}