% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{login}
\alias{login}
\title{Login & Create Security token.}
\usage{
login(email = "prompt", password = "prompt", simple = TRUE)
}
\arguments{
\item{email}{The SpatialMaps account email}

\item{password}{The SpatialMaps account password.}
}
\value{
returns list with various account information
}
\description{
The login function provides all information to interact with the SpatialMaps database. Its main functioanlity is to receive the JWT token required to upload MSnSets.
}
