\name{pRolocdata}
\alias{pRolocdata}

\title{
List of pRolocdata data sets
}

\description{
  This function returns a \code{dataframe} with the data sets name and
  description that are available within the \code{pRolocdata} package.
}

\usage{
pRolocdata()
}

\value{
  A \code{dataframe} with data sets names and descriptions.  
}

\references{
  See in the respective data sets' manual pages for references to
  publications. 
}

\author{
  Laurent Gatto <lg390@cam.ac.uk>
}


\examples{
pRolocdata()
}