% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pRolocFData}
\alias{pRolocFData}
\title{fData list aggregation.}
\usage{
pRolocFData(object)
}
\arguments{
\item{object}{The MSnSet.}
}
\description{
Helper function to reformat fData to DB suitable format.
}
