% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pRolocmetadata}
\alias{pRolocmetadata}
\alias{print.pRolocmetadata}
\title{Extract pRoloc metadata}
\usage{
pRolocmetadata(x)
}
\arguments{
\item{x}{A \code{pRolocdata} data.}
}
\value{
An instance of class \code{pRolocmetadata}.
}
\description{
Extracts relevant metadata from an \code{MSnSet} instance. See
\code{README.md} for a description and explanation of the metadata
fields.
}
\examples{
library("pRolocdata")
data(dunkley2006)
data(dunkley2006)
pRolocmetadata(dunkley2006)
}
\author{
Laurent Gatto
}

