% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pRolocMetaFrame}
\alias{pRolocMetaFrame}
\title{Metadata creation}
\usage{
pRolocMetaFrame(object, varName, token)
}
\arguments{
\item{object}{The MSnSet object.}

\item{token}{The JWT token created with the login() function.}

\item{varname}{The name of the object.}
}
\description{
Creates the /Meta data entry for SpatialMaps
}
