% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pRolocRawData}
\alias{pRolocRawData}
\title{s4 class to base64 binary}
\usage{
pRolocRawData(object)
}
\arguments{
\item{object}{The MsnSet object.}
}
\value{
A base64 encrypted MSnSet object list.
}
\description{
Transforms MSnSet object to binary base64 string. This step is required to keep store the s4 class without interfering with the JSON formatting requirements of firebase and to prevent any column reordering or other modifications.
}
