% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{resetPassword}
\alias{resetPassword}
\title{spatialMaps account password reset}
\usage{
resetPassword(email)
}
\arguments{
\item{email}{The SpatialMaps account email.}
}
\value{
Returns success or failure warning.
}
\description{
The password reset function in case the password was lost. The function triggers a password reset email sent to the input email.
}
