% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tokenCheck}
\alias{tokenCheck}
\title{Token verification}
\usage{
tokenCheck(token)
}
\arguments{
\item{token}{The JWT token created with the login() function.}
}
\value{
Reports the successfull transfer and outputs the random ID.
}
\description{
Verifies the authenticity of the generated token and returns the uniquer user identifier.
}
