% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{update}
\alias{update}
\title{Update existing SpatialMaps dataset.}
\usage{
update(dataset, name, token, randomKey)
}
\arguments{
\item{dataset}{A MSnSet object.}

\item{name}{Name of the MSnSet dataset.}

\item{token}{The JWT token created with the login() function.}

\item{randomKey}{The MSnSet random key. This key is provided either in on the 
SpatialMaps website or returned after each upload.}
}
\value{
Returns success message.
}
\description{
Other than the upload function, the update function takes a specified random key as path (other than creating a new one). The main value of the update function is not to create duplicates in the database.
}
