% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{upload}
\alias{upload}
\title{Upload Datasets from SpatialMaps}
\usage{
upload(dataset, name, token)
}
\arguments{
\item{dataset}{The proloc object.}

\item{name}{A string to add the name of the dataset.}

\item{token}{The JWT token created with the login() function.}
}
\value{
Reports the successfull transfer and outputs the random ID.
}
