\name{andy2011}
\alias{andy2011}
\alias{andy2011goCC}
\alias{andy2011hpa}


\docType{data}
\title{
  Data from a LOPIT experiment on Human Embryonic Kidney fibroblast cells
}

\description{
  This is a LOPIT dataset from a standard LOPIT experimental design on Human
  Embryonic Kidney (HEK293T) fibroblast cells. See below for more details. 
}

\usage{
data(andy2011)
data(andy2011goCC)
data(andy2011hpa)
}

\format{
  The data is an instance of class \code{MSnSet} from package \code{MSnbase}.
}

\details{
  This is a LOPIT experiment. Normalised intensities for 1371 proteins
  for eight iTRAQ 8-plex labelled fractions.
  This dataset was used in testing the phenotype discovery algorithm from 
  Breckels et al., \emph{The Effect of Organelle Discovery upon Sub-Cellular 
  Protein Localisation}, J Proteomics, 2013, 88:129-40, see \code{phenoDisco}. 
  New phenotype clusters identified from algorithm application are available as 
  \code{pd.2013} feature meta-data and the markers used as input for the analysis 
  are available as \code{markers} feature meta-data.

  The \code{andy2011goCC} instance contains binary assay data. Its
  columns represent GO CC terms that have been observed for the object's
  features. A 1 indicates that a GO term has been associated to a given
  feature (protein); a 0 means not such association was found in the GO
  ontology.
  
  The \code{andy2011hpa} instance contains binary assay data. Its
  columns represent subcellular locations that have been observed from 
  microscopy images from the Human Protein Atlas, for each protein. A 1 indicates that a subcellular term has been associated to a given
  feature (protein); a 0 means not such association was found. This matrix 
  of terms was generated from version 13, released on 11/06/2014 of the Human
  Protein Atlas. 
  
}

\source{
  The data was generated by A. Christoforou at the Cambridge Centre for 
  Proteomics. 

  \url{http://www.bio.cam.ac.uk/proteomics/}. 
}

\references{
  Breckels LM, Gatto L, Christoforou A, Groen AJ, Lilley KS and Trotter
  MWB. \emph{The Effect of Organelle Discovery upon Sub-Cellular Protein
  Localisation.} J Proteomics. 2013 Aug 2;88:129-40. doi:
  10.1016/j.jprot.2013.02.019. Epub 2013 Mar 21. PubMed PMID: 23523639
}

\examples{
data(andy2011)
andy2011 
pData(andy2011)
head(exprs(andy2011))
## Organelle marker proteins
table(fData(andy2011)$markers)
## PhenoDisco assignment results
table(fData(andy2011)$pd.2013)

data(andy2011goCC)
dim(andy2011goCC)
head(featureNames(andy2011goCC))
exprs(andy2011goCC)[1:10, 1:5]
}

\keyword{datasets}
