\name{andy2011}
\alias{andy2011}
\docType{data}
\title{
  Data from LOPIT experiment on Human Embryonic Kidney fibroblast cells
}

\description{
  This is a LOPIT dataset from a standard LOPIT experimental design on Human
  Embryonic Kidney (HEK293T) fibroblast cells. See below for more details. 
}

\usage{data(andy2011)}

\format{
  The data is an instance of class \code{MSnSet} from package \code{MSnbase}.
}

\details{
  This is a LOPIT experiment. Normalised intensities for 1371 proteins
  for eight iTRAQ 8-plex labelled fractions.
  This dataset was used in testing the phenotype discovery algorithm from 
  Breckels et al., \emph{The Effect of Organelle Discovery upon Sub-Cellular 
  Protein Localisation}, J Proteomics, \emph{In Press.}, see \code{phenoDisco}. 
  New phenotype clusters identified from algorithm application are available as 
  \code{pd.2013} feature meta-data and the markers used as input for the analysis 
  are available as \code{markers} feature meta-data.
}

\source{
  The data was generated by A. Christoforou at the Cambridge Centre for 
  Proteomics.
  \url{http://www.bio.cam.ac.uk/proteomics/}. 
}

\references{
  Breckels LM, Gatto L, Christoforou A, Groen AJ, Lilley KS and Trotter MWB. 
  \emph{The Effect of Organelle Discovery upon Sub-Cellular Protein Localisation}
   J Proteomics. \emph{In Press.}
  
}

\examples{
data(andy2011)
andy2011 
pData(andy2011)
head(exprs(andy2011))
## Organelle marker proteins
table(fData(andy2011)$markers)
## PhenoDisco assignment results
table(fData(andy2011)$pd.2013)
}

\keyword{datasets}
