\name{pRolocdata}
\alias{pRolocdata}

\title{
List of pRolocdata data sets
}

\description{
  This function lists the data sets available in \code{pRolocdata}
  package by calling \code{data(package = "pRolocdata")}. 
}

\usage{
pRolocdata()
}

\references{
  See in the respective data sets' manual pages for references to
  publications. 
}

\author{
  Laurent Gatto <lg390@cam.ac.uk>
}


\examples{
pRolocdata()
}